pokemon_mcp_project/requirements.txt
Django
djangorestframework
requests
python-dotenv
django-cors-headers
google-generativeai
